backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Speech Processing
openai-whisper
soundfile==0.12.1

# NLP & ML
transformers==4.36.0
sentencepiece==0.1.99
datasets==2.14.7
scikit-learn==1.3.2
numpy==1.26.4

# Deep Learning Framework
torch==2.1.1
torchaudio==2.1.1

# Text-to-Speech
gTTS==2.4.0
pyttsx3==2.90
pydub==0.25.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2

# Google Cloud
google-cloud-storage==2.14.0
google-cloud-functions-framework==3.5.0

